% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filters}
\alias{filters}
\alias{eq}
\alias{neq}
\alias{gt}
\alias{gte}
\alias{lt}
\alias{lte}
\alias{like}
\alias{ilike}
\alias{re_match}
\alias{re_imatch}
\alias{in_}
\alias{is_null}
\alias{is_true}
\alias{is_false}
\alias{not}
\title{Filter operators}
\usage{
eq(x)

neq(x)

gt(x)

gte(x)

lt(x)

lte(x)

like(pattern)

ilike(pattern)

re_match(pattern)

re_imatch(pattern)

in_(values)

is_null()

is_true()

is_false()

not(filter)
}
\arguments{
\item{x}{A single value to compare against. \code{Date} and \code{POSIXct} values are
formatted for the API; logicals become \code{true} and \code{false}.}

\item{pattern}{A pattern for \code{like()}, \code{ilike()}, \code{re_match()} and
\code{re_imatch()}.}

\item{values}{A vector of values for \code{in_()}.}

\item{filter}{A filter built by one of the other operators, to be negated.}
}
\value{
An object of class \code{tg_filter}.
}
\description{
Constructors for the comparison operators the 'PostgREST' services behind the
TransfereGov APIs accept. Pass them as named arguments to \code{\link[=tg_get]{tg_get()}} or
\code{\link[=tg_count]{tg_count()}}, where the name is the column being filtered.
}
\details{
A bare value is shorthand for \code{eq()}, and a bare vector of length greater
than one is shorthand for \code{in_()}, so \code{aa_ano_plano_acao = 2024} and
\code{aa_ano_plano_acao = c(2024, 2025)} both work. Pass a list of operators to
apply several conditions to the same column, which the API combines with AND:
\code{dt_inicio_vigencia = list(gte("2024-01-01"), lt("2025-01-01"))}.

In \code{like()} and \code{ilike()} the wildcard may be written as \code{*} or \verb{\%}. In
\code{re_match()} and \code{re_imatch()} the operand is a POSIX regular expression.
}
\examples{
gte(2024)
in_(c("PE", "PB"))
not(is_null())

if (interactive()) {
  tg_get("ted", "plano_acao", aa_ano_plano_acao = gte(2024))
}
}
\seealso{
Other filters:
\code{\link[=tg_operators]{tg_operators()}}
}
\concept{filters}
