% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{tg_cache_clear}
\alias{tg_cache_clear}
\alias{tg_cache_limpar}
\title{Delete cached responses}
\usage{
tg_cache_clear()

tg_cache_limpar()
}
\value{
The number of files removed, invisibly.
}
\description{
Delete cached responses
}
\examples{
tg_cache_clear()
}
\seealso{
Other cache:
\code{\link[=tg_cache_dir]{tg_cache_dir()}}
}
\concept{cache}
