% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{tg_operators}
\alias{tg_operators}
\alias{tg_operadores}
\title{List the available filter operators}
\usage{
tg_operators()

tg_operadores()
}
\value{
A tibble with the exported operator, the 'PostgREST' operator it
sends, and what it means.
}
\description{
List the available filter operators
}
\examples{
tg_operators()
}
\seealso{
Other filters:
\code{\link{filters}}
}
\concept{filters}
