% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{module_shortcuts}
\alias{module_shortcuts}
\alias{tg_ted}
\alias{tg_fundo_a_fundo}
\alias{tg_transferencias_especiais}
\title{Query a single module}
\usage{
tg_ted(table, ...)

tg_fundo_a_fundo(table, ...)

tg_transferencias_especiais(table, ...)
}
\arguments{
\item{table}{A table name from \code{\link[=tg_tables]{tg_tables()}} for that module.}

\item{...}{Passed to \code{\link[=tg_get]{tg_get()}}: filters, and any of its \code{.}-prefixed
arguments.}
}
\value{
A tibble, as \code{\link[=tg_get]{tg_get()}} returns.
}
\description{
Thin wrappers over \code{\link[=tg_get]{tg_get()}} with the module fixed, for code that stays
within one API.
}
\examples{
if (interactive()) {
  tg_ted("plano_acao", .limit = 10)
  tg_fundo_a_fundo("programa", .limit = 10)
  tg_transferencias_especiais("programa_especial")
}
}
\seealso{
Other queries:
\code{\link[=tg_count]{tg_count()}},
\code{\link[=tg_get]{tg_get()}},
\code{\link[=tg_metadata]{tg_metadata()}}
}
\concept{queries}
